# distsql: refactor hashJoiner to implement RowSource